refactor: menu item global style
