+++ packages/components/menu/src/MenuItem/MenuItem.tsx
@@ -76,7 +76,7 @@ function _MenuItem<E extends React.ElementType = typeof MENU_ITEM_DEFAULT_TAG>(
       {...otherProps}
       {...getMenuItemProps(otherProps)}
       disabled={isDisabled ?? props.disabled}
-      className={cx(styles.root, className)}
+      className={cx(styles, className)}
       data-test-id={itemTestId}
       ref={mergeRefs(itemRef, ref)}
       tabIndex={-1}
