+++ packages/core/src/index.ts
@@ -30,3 +30,4 @@ export { useControllableState, useId, useImageLoaded } from './hooks';
 export type { UseControllableStateProps } from './hooks';
 export { mergeRefs } from './utils/mergeRefs';
 export { getEntityStatusStyles } from './utils/getEntityStatusStyles';
+export { getMenuItemStyles } from './utils/getMenuItem.styles';
